Revert "Entity"
