+++ api/src/main/java/com/depromeet/articlereminder/domain/AlramState.java
@@ -1,0 +1,5 @@
+package com.depromeet.articlereminder.domain;
+
+public enum AlramState {
+    ON,OFF
+}
