+++ api/src/main/java/com/depromeet/articlereminder/domain/Hashtag.java
@@ -1,0 +1,23 @@
+package com.depromeet.articlereminder.domain;
+
+import org.springframework.data.annotation.CreatedDate;
+
+import javax.persistence.Column;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import java.time.LocalDateTime;
+
+public class Hashtag {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "hashtag_id")
+    private Long id;
+
+    private String name;
+
+    @CreatedDate
+    private LocalDateTime createdAt;
+
+}
