+++ api/src/main/java/com/depromeet/articlereminder/domain/Link.java
@@ -1,0 +1,17 @@
+package com.depromeet.articlereminder.domain;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
+public class Link {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+//    private String ;
+
+}
