--- api/src/main/java/com/depromeet/articlereminder/domain/alarm/AlarmState.java
@@ -1,5 +1,0 @@
-package com.depromeet.articlereminder.domain.alarm;
-
-public enum AlarmState {
-    ON,OFF
-}
