--- api/src/main/java/com/depromeet/articlereminder/domain/alarm/AlarmStatus.java
@@ -1,6 +1,0 @@
-package com.depromeet.articlereminder.domain.alarm;
-
-public enum AlarmStatus {
-    ENABLED,
-    DISABLED
-}
