+++ api/src/main/java/com/depromeet/articlereminder/domain/alarm/PushAlarm.java
@@ -1,5 +1,7 @@
 package com.depromeet.articlereminder.domain.alarm;
 
+import com.depromeet.articlereminder.domain.AlramState;
+import com.depromeet.articlereminder.domain.RepeatedDate;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -11,5 +13,5 @@ import javax.persistence.Entity;
 public class PushAlarm extends Alarm {
 
     private RepeatedDate repeatedDate;
-    private AlarmState alramState;
+    private AlramState alramState;
 }
