--- api/src/main/java/com/depromeet/articlereminder/domain/hashtag/Hashtag.java
@@ -1,26 +1,0 @@
-package com.depromeet.articlereminder.domain.hashtag;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Hashtag {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "hashtag_id")
-    private Long id; // 해시태그 id
-
-    private String name; // 해시태그명
-
-    @CreatedDate
-    private LocalDateTime createdAt; // 생성 일시
-
-}
