--- api/src/main/java/com/depromeet/articlereminder/domain/link/LinkStatus.java
@@ -1,6 +1,0 @@
-package com.depromeet.articlereminder.domain.link;
-
-public enum LinkStatus {
-    UNREAD,
-    READ
-}
