--- api/src/main/java/com/depromeet/articlereminder/domain/ranking/Ranking.java
@@ -1,38 +1,0 @@
-package com.depromeet.articlereminder.domain.ranking;
-
-
-import com.depromeet.articlereminder.domain.member.Member;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-
-import javax.persistence.*;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-
-import static javax.persistence.FetchType.*;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Ranking {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "rank_id")
-    private Long id;
-
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member; // 사용자
-
-    private Integer rank; // 순위
-
-    private LocalDate baseDt; // 기준 일자
-
-    private Integer count; // 아티클 읽은 개수
-
-    @CreatedDate
-    private LocalDateTime loadDtm; // 생성 일시
-}
