--- api/src/main/java/com/depromeet/articlereminder/domain/ranking/SeasonalCount.java
@@ -1,31 +1,0 @@
-package com.depromeet.articlereminder.domain.ranking;
-
-
-import com.depromeet.articlereminder.domain.member.Member;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-import java.time.LocalDate;
-
-import static javax.persistence.FetchType.LAZY;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class SeasonalCount {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @JoinColumn(name = "seasonal_id")
-    private Long id;
-
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    private LocalDate gainDt; // 포인트 획득 일자
-
-    private Integer count; // 읽은 아티클 갯수
-}
